Prepify/requirements.txt
# Core Django packages
Django>=4.2.0,<5.0
djangorestframework>=3.14.0

# Audio processing - Let pip resolve compatible versions
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.21.0

# AI/ML packages - Use latest compatible versions
sentence-transformers>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Grammar checking
language-tool-python>=2.7.0

# Additional utilities
Pillow>=9.0.0
python-dotenv>=1.0.0

# Development tools (optional)
django-extensions>=3.2.0